Redo: Changes variables in src/tree.c to short

+++ src/tree.h
@@ -10,8 +10,8 @@
  * be initialized with real board and values
  */
 Node *initNode(void);
-// void createTree(Node *root, const ssize_t depth);
-Node *createChild(Node *parent, const size_t child);
+// void createTree(Node *root, const short depth);
+Node *createChild(Node *parent, const short child);
 void createChildren(Node *parent);
 // void printTree(const Node *root);
